pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "design-agents"
version = "0.2.0"
description = "VNext skill-governed agent runtime with src-first registry, context, and surface assembly."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["agents*", "context*", "governance*", "llm*", "runtime*", "schemas*", "shared*", "skills*", "storage*", "tools*"]

[tool.setuptools.package-data]
agents = ["specs/*.agent.yaml"]
context = ["templates/*.md"]
skills = ["**/*.md"]

